refactor(import/svg): move the runtime-Paint projection off the IR
The IRSVG types in cg/svg.rs carried the SVG->runtime-Paint projection
(into_paint_with_opacity + gradient helpers) as inherent impls. That
projection encodes painter conventions — baked fill/stroke opacity,
gradient transforms normalized into the painter's UV [0,1] space, the
radial focal-point drop — i.e. adapter policy, not SVG-spec semantics.

Move it verbatim to import/svg/paint.rs as free functions
(fill_paint / stroke_paint + private helpers). Free functions rather
than From/inherent impls: after the planned cg/ crate extraction,
impls over those types from this crate would violate the orphan rule.
The From<SVGGradientSpreadMethod> for TileMode impl becomes the
private tile_mode_from_spread — the Pad/Reflect/Repeat -> tile-mode
mapping is the painter's interpretation, so it moves with the
projection.

cg/svg.rs is now vocabulary-only (zero runtime Paint references).
The six pack.rs call sites rewire mechanically; helper bodies are
byte-identical (float-sensitive gradient math untouched).

Part of the SVG sink inversion (gridaco/nothing#29), under the seam
program gridaco/nothing#27.
